Revert "[NFTIO-3756] deps update (#1873)" (#1878)
This reverts commit 84370b5e78629cbb24209cd7c7fa15cfb392b6c3.

+++ src/pallet/marketplace/processors/listing-filled.ts
@@ -53,7 +53,7 @@ export async function listingFilled(
         id: `${listingId}-${item.id}`,
         amount: event.amountFilled,
         buyer,
-        price: 'price' in event ? event.price : listing.highestPrice,
+        price: 'price' in event ? (event.price as bigint) : listing.highestPrice,
         listing,
         createdAt: new Date(block.timestamp ?? 0),
     })
@@ -145,7 +145,7 @@ export async function listingFilled(
                 id: buyer.id,
             },
             amountFilled: event.amountFilled,
-            price: 'price' in event ? event.price : listing.highestPrice,
+            price: 'price' in event ? (event.price as bigint) : listing.highestPrice,
             amountRemaining: event.amountRemaining,
             protocolFee: event.protocolFee,
             royalty: event.royalty,
